+++ NEWS.md
@@ -1,3 +1,12 @@
+# eventPred 0.1.2
+
+- add mean and variance to prediction output
+- update the BIC weight for model averaging
+- add more details for model fit parameters 
+- add day 1 to enrollment plot
+- allow prior piecewise Poisson enrollment and piecewise exponential event or dropout models to have additional cut points beyond the observed data range
+- update internal data sets
+
 # eventPred 0.1.1
 
 - add stage and to_predict information in getPrediction output